refactor: Registry 클래스들을 프레임워크 패키지로 이동
